third_party/WebKit/LayoutTests/platform/mac/virtual/reporting/external/wpt/content-security-policy/reporting/reporting-api-report-only-sends-reports-on-violation.https.sub-expected.txt

This is a testharness.js-based test.
PASS Test that image does not load
PASS Event is fired
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
Harness: the test ran to completion.
